bluejay feedback: rely on serverless environments for controlling bootstrap
This introduces SENTRY_DSN environment as a stub, as well as adds
in an XRAY_ENABLED flag, to control when we load xray.

This is needed, as XRAY does not run well while testing, and can't
be reliably turned off with fixtures.

The behaviour is this:
- if SENTRY_DSN is not set, the Sentry client will internally not transmit
  anything.
- if XRAY_ENABLED is not 'true', no patching will be performed, and the sdk
  will be disabled.

SENTRY_DSN must be configured by the user.
XRAY_ENABLED will be set if lambda tracing is enabled (default)