Use the Generated Role from Aviary
We have a problem where Serverless wants to generate a role, however there isn't
a great way to confine it.
This prevents us from running CD, as that's a lot of access to grant a 3rd party.

We now generate a Role as part of Aviary using the 'serverless' Atom.

This also has the effect of making it easier to define custom permissions relating to
Aviary managed resources, such as SSM Parameters, or SQS subscriptions